GOV-166 Implementando la lógica de unirse a las comunidades

+++ src/main/java/com/upao/govench/govench/api/CommunityController.java
@@ -4,19 +4,24 @@ import com.upao.govench.govench.model.entity.Community;
 import com.upao.govench.govench.model.entity.User;
 import com.upao.govench.govench.service.CommunityService;
 import com.upao.govench.govench.service.EncryptionService;
+import com.upao.govench.govench.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.nio.file.AccessDeniedException;
 import java.util.List;
 
-@RequestMapping("api/v1/{encodedUserId}/community")
+@RequestMapping("/{encodedUserId}/community")
 @RestController
 public class CommunityController {
     @Autowired
     private EncryptionService encryptionService; // Inyección del servicio de encriptación
     @Autowired
     private CommunityService CommunityService; // Inyección del servicio de Community
+    @Autowired
+    private UserService UserService;
 
     @GetMapping
     public List<Community> getCommunitiesByUser(@PathVariable("encodedUserId") String encodedUserId) throws Exception {
@@ -37,9 +42,9 @@ public class CommunityController {
         Community existingCommunity = CommunityService.findById(id);
 
         // Verificar que el usuario es el propietario de la comunidad
-       // if (existingCommunity.getOwner().getId() != userId) {
-       //     throw new AccessDeniedException("No tienes permiso para eliminar esta comunidad");
-        //}
+       if (existingCommunity.getOwner().getId() != userId) {
+           throw new AccessDeniedException("No tienes permiso para eliminar esta comunidad");
+        }
         CommunityService.deleteById(id);
     }
 
@@ -48,18 +53,28 @@ public class CommunityController {
         return CommunityService.findById(id);
     }
     @PutMapping("/{id}")
-    public Community updateCommunity(@PathVariable("encodedUserId") String encodedUserId, @PathVariable("id") int id, @RequestBody Community community) throws Exception {
+    public ResponseEntity<String> updateCommunity(@PathVariable("encodedUserId") String encodedUserId, @PathVariable("id") int id, @RequestBody Community community) throws Exception
+    {
         int userId = Integer.parseInt(encryptionService.decrypt(encodedUserId));
 
         // Obtener la comunidad por su ID
         Community existingCommunity = CommunityService.findById(id);
 
         // Verificar que el usuario es el propietario de la comunidad
-         // if (existingCommunity.getOwner().getId() != userId) {
-        //   throw new AccessDeniedException("No tienes permiso para modificar esta comunidad");
-        // }
+          if (existingCommunity.getOwner().getId() != userId) {
+          throw new AccessDeniedException("No tienes permiso para modificar esta comunidad");
+        }
         
         // Si es el propietario, permite la actualización
-        return CommunityService.update(community, id);
+        CommunityService.update(community, id);
+        return new ResponseEntity<>("Comunidad editada con éxito", HttpStatus.ACCEPTED);
+    }
+    @PostMapping("/create")
+    public ResponseEntity<String> createCommunity(@PathVariable("encodedUserId") String encodedUserId,@RequestBody Community community) throws Exception {
+        int userId = Integer.parseInt(encryptionService.decrypt(encodedUserId));
+        User owner=UserService.getUserbyId(userId);
+        CommunityService.save(community,owner);
+        return new ResponseEntity<>("Comunidad creada con éxito", HttpStatus.CREATED);
+
     }
 }
